mountain (refactor): using frames for chairs
